dashboard/requirements.txt
pandas
boto3
altair
pymssql
python-dotenv
streamlit